Support custom SSH config file for SSH connection
SSH Docker host connections support specifying an SSH username,
hostname, port, and socket path with `ssh://…`. Furthermore, the default
user SSH config file located at `~/.ssh/config` is respected.

However, it would be useful to specify a path to a custom SSH config on
the CLI. Use cases are per-project SSH configs or SSH configs generated
on the fly (e.g., by `vagrant ssh-config`).

With this commit, a path to a custom SSH config can be specified as part
of the SSH Docker host URL.

This closes #1301. Related issues: #1463, #2539.

Signed-off-by: Benjamin Fischer <benjamin.fischer@evolutics.info>