add support for DOCKER_CUSTOM_HEADERS env-var
This environment variable allows for setting additional headers
to be sent by the client. Headers set through this environment
variable are added to headers set through the config-file (through
the HttpHeaders field).

This environment variable can be used in situations where headers
must be set for a specific invocation of the CLI, but should not
be set by default, and therefore cannot be set in the config-file.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 0d6b805256207aa5601956fa83ed94951b9f71b6)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>