auth: add support for oauth device-code login
This commit adds support for the oauth [device-code](https://auth0.com/docs/get-started/authentication-and-authorization-flow/device-authorization-flow)
login flow when authenticating against the official registry.

This is achieved by adding `cli/internal/oauth`, which contains code to manage
interacting with the Docker OAuth tenant (`login.docker.com`), including launching
the device-code flow, refreshing access using the refresh-token, and logging out.

The `OAuthManager` introduced here is also made available through the `command.Cli`
interface method `OAuthManager()`.

In order to manage storage and retrieval of the oauth credentials, as well as ensuring
that the access token retrieved from the credentials store is valid when accessed, this
commit also introduces a new credential store wrapper, `OAuthStore`, which defers storage
to an underlying store (such as the file or native/keychain/pass/wincred store) and
gets a new access token using the refresh token when accessed.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>